Moved /shape folder to /core/shape
