+++ src/core/shape/bounding-sphere.js
@@ -1,5 +1,5 @@
-import { Debug } from '../core/debug.js';
-import { Vec3 } from '../core/math/vec3.js';
+import { Debug } from '../debug.js';
+import { Vec3 } from '../math/vec3.js';
 
 /** @typedef {import('./ray.js').Ray} Ray */
