+++ src/core/shape/frustum.js
@@ -1,7 +1,7 @@
-import { Vec3 } from '../core/math/vec3.js';
-import { PROJECTION_PERSPECTIVE } from '../scene/constants.js';
+import { Vec3 } from '../math/vec3.js';
+import { PROJECTION_PERSPECTIVE } from '../../scene/constants.js';
 
-/** @typedef {import('../core/math/mat4.js').Mat4} Mat4 */
+/** @typedef {import('../math/mat4.js').Mat4} Mat4 */
 /** @typedef {import('./bounding-sphere.js').BoundingSphere} BoundingSphere */
 
 const _frustumPoints = [new Vec3(), new Vec3(), new Vec3(), new Vec3(), new Vec3(), new Vec3(), new Vec3(), new Vec3()];
