+++ src/core/shape/oriented-box.js
@@ -1,6 +1,6 @@
-import { Debug } from '../core/debug.js';
-import { Mat4 } from '../core/math/mat4.js';
-import { Vec3 } from '../core/math/vec3.js';
+import { Debug } from '../debug.js';
+import { Mat4 } from '../math/mat4.js';
+import { Vec3 } from '../math/vec3.js';
 
 import { BoundingBox } from './bounding-box.js';
 import { BoundingSphere } from './bounding-sphere.js';
