+++ src/framework/components/camera/component.js
@@ -12,7 +12,7 @@ import { Debug } from '../../../core/debug.js';
 /** @typedef {import('../../../core/math/vec3.js').Vec3} Vec3 */
 /** @typedef {import('../../../core/math/vec4.js').Vec4} Vec4 */
 /** @typedef {import('../../../scene/layer.js').Layer} Layer */
-/** @typedef {import('../../../shape/frustum.js').Frustum} Frustum */
+/** @typedef {import('../../../core/shape/frustum.js').Frustum} Frustum */
 /** @typedef {import('../../../xr/xr-manager.js').XrErrorCallback} XrErrorCallback */
 /** @typedef {import('../../entity.js').Entity} Entity */
 /** @typedef {import('./system.js').CameraComponentSystem} CameraComponentSystem */
