+++ src/framework/components/model/system.js
@@ -2,7 +2,7 @@ import { extend } from '../../../core/core.js';
 
 import { Vec3 } from '../../../core/math/vec3.js';
 
-import { BoundingBox } from '../../../shape/bounding-box.js';
+import { BoundingBox } from '../../../core/shape/bounding-box.js';
 
 import { getDefaultMaterial } from '../../../scene/materials/default-material.js';
