+++ src/framework/components/render/system.js
@@ -1,6 +1,6 @@
 import { Vec3 } from '../../../core/math/vec3.js';
 
-import { BoundingBox } from '../../../shape/bounding-box.js';
+import { BoundingBox } from '../../../core/shape/bounding-box.js';
 
 import { getDefaultMaterial } from '../../../scene/materials/default-material.js';
