+++ src/input/element-input.js
@@ -3,7 +3,7 @@ import { platform } from '../core/platform.js';
 import { Vec3 } from '../core/math/vec3.js';
 import { Vec4 } from '../core/math/vec4.js';
 
-import { Ray } from '../shape/ray.js';
+import { Ray } from '../core/shape/ray.js';
 
 import { getApplication } from '../framework/globals.js';
