+++ src/resources/parser/json-model.js
@@ -3,7 +3,7 @@ import { Debug } from '../../core/debug.js';
 import { Mat4 } from '../../core/math/mat4.js';
 import { Vec3 } from '../../core/math/vec3.js';
 
-import { BoundingBox } from '../../shape/bounding-box.js';
+import { BoundingBox } from '../../core/shape/bounding-box.js';
 
 import {
     INDEXFORMAT_UINT16, INDEXFORMAT_UINT32,
