+++ src/xr/xr-input-source.js
@@ -4,7 +4,7 @@ import { Mat4 } from '../core/math/mat4.js';
 import { Quat } from '../core/math/quat.js';
 import { Vec3 } from '../core/math/vec3.js';
 
-import { Ray } from '../shape/ray.js';
+import { Ray } from '../core/shape/ray.js';
 
 import { XrHand } from './xr-hand.js';
